Move PsiResolutionNext into PsiResolutionStrategy
