--- kotlin-analysis-api/src/main/kotlin/com/google/devtools/ksp/common/impl/PsiResolutionNext.kt
@@ -1,991 +1,0 @@
-/*
- * Copyright 2026 Google LLC
- * Copyright 2010-2026 JetBrains s.r.o. and Kotlin Programming Language contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.devtools.ksp.common.impl
-
-import com.google.devtools.ksp.InternalKSPException
-import com.google.devtools.ksp.common.visitor.CollectAnnotatedSymbolsPsiVisitor
-import com.google.devtools.ksp.containingFile
-import com.google.devtools.ksp.impl.FileCache
-import com.google.devtools.ksp.impl.symbol.kotlin.KSClassDeclarationImpl
-import com.google.devtools.ksp.impl.symbol.kotlin.KSFileImpl
-import com.google.devtools.ksp.impl.symbol.kotlin.KSFunctionDeclarationImpl
-import com.google.devtools.ksp.impl.symbol.kotlin.KSPropertyDeclarationJavaImpl
-import com.google.devtools.ksp.impl.symbol.kotlin.Restorable
-import com.google.devtools.ksp.impl.symbol.kotlin.analyze
-import com.google.devtools.ksp.impl.symbol.kotlin.getFqn
-import com.google.devtools.ksp.impl.symbol.kotlin.getGeneratedProperty
-import com.google.devtools.ksp.impl.symbol.kotlin.getter
-import com.google.devtools.ksp.impl.symbol.kotlin.psi
-import com.google.devtools.ksp.impl.symbol.kotlin.setter
-import com.google.devtools.ksp.impl.symbol.kotlin.toKSAnnotated
-import com.google.devtools.ksp.impl.symbol.kotlin.toKSAnnotationUseSiteTarget
-import com.google.devtools.ksp.impl.symbol.kotlin.toKSClassDeclaration
-import com.google.devtools.ksp.impl.symbol.kotlin.toKSFile
-import com.google.devtools.ksp.impl.symbol.kotlin.toKSFunctionDeclaration
-import com.google.devtools.ksp.impl.symbol.kotlin.toKSPropertyDeclaration
-import com.google.devtools.ksp.impl.symbol.kotlin.toKtClassSymbol
-import com.google.devtools.ksp.impl.symbol.kotlin.toLocation
-import com.google.devtools.ksp.processing.SymbolProcessor
-import com.google.devtools.ksp.symbol.AnnotationUseSiteTarget
-import com.google.devtools.ksp.symbol.KSAnnotated
-import com.google.devtools.ksp.symbol.KSBackingField
-import com.google.devtools.ksp.symbol.KSFile
-import com.google.devtools.ksp.symbol.KSFunctionDeclaration
-import com.google.devtools.ksp.symbol.KSPropertyDeclaration
-import com.google.devtools.ksp.symbol.KSTypeParameter
-import com.google.devtools.ksp.symbol.KSValueParameter
-import com.google.devtools.ksp.symbol.NonExistLocation
-import com.intellij.psi.PsiAnnotation
-import com.intellij.psi.PsiClass
-import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiField
-import com.intellij.psi.PsiMember
-import com.intellij.psi.PsiMethod
-import com.intellij.psi.PsiModifierListOwner
-import com.intellij.psi.PsiParameter
-import com.intellij.psi.PsiTypeParameter
-import com.intellij.psi.PsiTypeParameterList
-import com.intellij.util.containers.addIfNotNull
-import org.jetbrains.kotlin.analysis.api.annotations.KaAnnotation
-import org.jetbrains.kotlin.analysis.api.annotations.KaAnnotationValue
-import org.jetbrains.kotlin.analysis.api.resolution.KaAnnotationCall
-import org.jetbrains.kotlin.analysis.api.resolution.successfulCallOrNull
-import org.jetbrains.kotlin.analysis.api.resolution.symbol
-import org.jetbrains.kotlin.analysis.api.symbols.KaCallableSymbol
-import org.jetbrains.kotlin.analysis.api.symbols.KaClassKind
-import org.jetbrains.kotlin.analysis.api.symbols.KaClassSymbol
-import org.jetbrains.kotlin.analysis.api.symbols.KaTypeAliasSymbol
-import org.jetbrains.kotlin.analysis.api.types.symbol
-import org.jetbrains.kotlin.name.CallableId
-import org.jetbrains.kotlin.name.ClassId
-import org.jetbrains.kotlin.name.FqName
-import org.jetbrains.kotlin.name.Name
-import org.jetbrains.kotlin.psi.KtAnnotated
-import org.jetbrains.kotlin.psi.KtAnnotationEntry
-import org.jetbrains.kotlin.psi.KtDeclaration
-import org.jetbrains.kotlin.psi.KtFile
-import org.jetbrains.kotlin.psi.KtTypeReference
-import org.jetbrains.kotlin.psi.psiUtil.parameterIndex
-import org.jetbrains.kotlin.utils.addToStdlib.flatGroupBy
-
-/**
- * This is a helper class for [PsiResolutionStrategy] which supports upcoming language features.
- * The shape of [PsiResolutionNext] is intentionally kept very similar to that of [PsiResolutionStrategy] to allow
- * easy merging of the two classes.
- *
- * It is intentional that it does **not** implement [AnnotationResolutionStrategy], since it should only
- * be used in a specific scenario where [PsiResolutionStrategy] calls it.
- */
-class PsiResolutionNext(
-    val newKSFiles: List<KSFile>,
-    val deferredSymbols: Map<SymbolProcessor, List<Restorable>>
-) {
-
-    private val targetFieldFix = AnnotationDeclarationTargetFieldFix()
-
-    /**
-     * Returns all symbols annotated with [annotationName].
-     */
-    fun getSymbolsWithAnnotation(annotationName: String): Sequence<KSAnnotated> =
-        // getAnnotatedSymbols(annotationName)
-        throw InternalKSPException(
-            "Unexpected unreachable function call. " +
-                "PsiResolutionNext.getSymbolsWithAnnotation($annotationName) should not be called. " +
-                "It is not fully implemented yet",
-            NonExistLocation,
-            javaClass
-        )
-
-    /**
-     * Returns the symbols annotated with the fully qualified name [annotationName].
-     */
-    private fun getAnnotatedSymbols(annotationName: String): Sequence<KSAnnotated> =
-        getAnnotatedJavaSymbols(annotationName) +
-            getAnnotatedKotlinSymbols(annotationName) +
-            getRestoredSymbols(annotationName)
-
-    /**
-     * Returns the Kotlin symbols annotated with the fully qualified name [annotationName].
-     */
-    private fun getAnnotatedKotlinSymbols(annotationName: String): Sequence<KSAnnotated> {
-        val newKotlinSymbols = annotatedKotlinElementsByFullyQualifiedName[annotationName]?.value ?: emptyList()
-        return newKotlinSymbols.asSequence()
-    }
-
-    /**
-     * Returns the Java symbols annotated with the fully qualified name [annotationName].
-     */
-    private fun getAnnotatedJavaSymbols(annotationName: String): Sequence<KSAnnotated> =
-        annotatedJavaElementsByFullyQualifiedName.getOrPut(annotationName) {
-            resolveJavaAnnotationByShortName(annotationName)
-        }.asSequence()
-
-    /**
-     * Returns the [deferredSymbols] that are annotated with [annotationName].
-     */
-    private fun getRestoredSymbols(annotationName: String): Sequence<KSAnnotated> =
-        deferredSymbolsGroupedByAnnotations[annotationName]?.asSequence() ?: emptySequence()
-
-    /**
-     * [deferredSymbols] grouped by their fully qualified annotation names.
-     *
-     * See [annotatedKotlinElementsByFullyQualifiedName] for an example of the grouping.
-     */
-    private val deferredSymbolsGroupedByAnnotations: Map<String, Collection<KSAnnotated>> by lazy {
-        deferredSymbols.values.flatten().mapNotNull { it.restore() }.distinct()
-            .flatGroupBy { sym -> sym.annotations.mapNotNull { it.getFqn() }.toSet() }
-    }
-
-    /**
-     * Returns the Java symbols annotated with the fully qualified name [annotationName].
-     * The function only considers the subset of annotations in the program that share the same
-     * short name / unqualified name.
-     */
-    private fun resolveJavaAnnotationByShortName(annotationName: String): Collection<KSAnnotated> {
-        val annotationShortName = annotationName.substringAfterLast('.')
-        return annotatedJavaElementsByShortName[annotationShortName]?.filter { element ->
-            val annotationsForElement = fullyQualifiedJavaAnnotationNamesByElements[element] ?: emptyMap()
-            annotationsForElement[annotationShortName]?.value?.contains(annotationName) ?: false
-        }?.flatMap {
-            it.resolveToKSAnnotated()
-        } ?: emptyList()
-    }
-
-    /**
-     * All annotated [PsiElement]s in [newKSFiles] from both Kotlin and Java sources.
-     */
-    private val annotatedPsiElements: Collection<PsiElement> by lazy {
-        newKSFiles.flatMap { collectAnnotatedPsiElementsIn(it) }
-    }
-
-    /**
-     * Returns all [PsiElement]s that are annotated in [file] except for declarations
-     * in function bodies or property accessors.
-     */
-    private fun collectAnnotatedPsiElementsIn(file: KSFile): Collection<PsiElement> {
-        val visitor = CollectAnnotatedSymbolsPsiVisitor()
-        file.psi?.accept(visitor)
-        return visitor.result
-    }
-
-    /**
-     * All annotated Java elements.
-     */
-    private val annotatedJavaElements: List<PsiModifierListOwner> by lazy {
-        annotatedPsiElements.filterIsInstance<PsiModifierListOwner>()
-    }
-
-    /**
-     * Groups [annotatedJavaElements] by the short names of their annotations.
-     */
-    private val annotatedJavaElementsByShortName: Map<String, Collection<PsiElement>> by lazy {
-        annotatedJavaElements.flatGroupBy { element ->
-            // N.B.: deduplicate by converting to a set, since
-            // otherwise it will add `element` more than once to the list of values
-            element.annotations.map { it.shortName }.toSet()
-        }
-    }
-
-    /**
-     * A map from Java elements to their annotations' fully qualified names, indexed by short name.
-     */
-    private val fullyQualifiedJavaAnnotationNamesByElements: Map<PsiElement, Map<String, Lazy<Set<String>>>> by lazy {
-        annotatedJavaElements.associate { element ->
-            // N.B.: Performance: Lazily compute the set of qualified names to avoid resolution until requested
-            //  for a specific annotation.
-            element to buildMap<String, Lazy<MutableSet<String>>> {
-                element.annotations.forEach { anno ->
-                    val lazyFullyQualifiedNames = this[anno.shortName]
-                    if (lazyFullyQualifiedNames == null) {
-                        this[anno.shortName] = lazy {
-                            mutableSetOf(
-                                anno.qualifiedName
-                                    ?: throw InternalKSPException(
-                                        "Unexpected unqualified name",
-                                        anno.toLocation(),
-                                        anno.javaClass
-                                    )
-                            )
-                        }
-                    } else {
-                        this[anno.shortName] = lazy {
-                            lazyFullyQualifiedNames.value.add(
-                                anno.qualifiedName
-                                    ?: throw InternalKSPException(
-                                        "Unexpected unqualified name",
-                                        anno.toLocation(),
-                                        anno.javaClass
-                                    )
-                            )
-                            lazyFullyQualifiedNames.value
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * Groups all annotated Java symbols by their fully qualified annotation names.
-     *
-     * @see [annotatedKotlinElementsByFullyQualifiedName] for further explanation.
-     */
-    private val annotatedJavaElementsByFullyQualifiedName: MutableMap<String, Collection<KSAnnotated>> =
-        mutableMapOf()
-
-    /**
-     * All annotated Kotlin elements.
-     */
-    private val annotatedKotlinElements: List<KtAnnotated> by lazy {
-        annotatedPsiElements.filterIsInstance<KtAnnotated>()
-    }
-
-    /**
-     * Groups all annotated Kotlin symbols by their fully qualified annotation names.
-     *
-     * In other words, if `fun1, fun2, fun3` are annotated as follows:
-     * ```
-     * @Annotation1 fun fun1
-     * @Annotation1 @Annotation2 fun fun2
-     * @Annotation2 @Annotation3 fun fun3
-     * ```
-     * Then the result is
-     *
-     * ```
-     * "Annotation1" -> [fun1, fun2]
-     * "Annotation2" -> [fun2, fun3]
-     * "Annotation3" -> [fun3]
-     * ```
-     */
-    private val annotatedKotlinElementsByFullyQualifiedName: Map<String, Lazy<List<KSAnnotated>>> by lazy {
-        buildMap {
-            val fileCaches = mutableMapOf<KtFile, FileCache>()
-            annotatedKotlinElements.forEach { annotated ->
-                val file = annotated.containingKtFile
-                annotated.annotationEntries.forEach { annotationEntry ->
-                    // TODO: Cache (file, annotation.shortName) -> classId. Shadowing is a problem
-                    //  so we should only cache for known unique declarations in the file.
-                    //  See test shadowingAnnotations.kt
-                    val fileCache = fileCaches.getOrPut(file) { FileCache(file) }
-                    // N.B.: Skip nullable qualified names without error. This is what the AA-based implementation does.
-                    //   For now, this is the intended behavior. If it is changed in the future, then it is an API change.
-                    val classId = resolveAnnotationEntryClassId(annotationEntry, fileCache) ?: return@forEach
-                    targetFieldFix.recordClassId(annotationEntry, classId)
-                    val fqn = classId.asFqNameString()
-                    getOrPut(fqn, ::mutableListOf)
-                        .add(annotated to annotationEntry)
-                }
-            }
-        }.mapValues { entry ->
-            lazy {
-                entry.value.flatMap { (annotated, annotationEntry) ->
-                    annotated.resolveToKSAnnotated(annotationEntry)
-                }.distinct()
-            }
-        }
-    }
-
-    /**
-     * Resolves the [ClassId] for a given [KtAnnotationEntry].
-     *
-     * It first attempts a [fastResolveClassId] and falls back to [slowResolveClassId] if that fails.
-     */
-    private fun resolveAnnotationEntryClassId(annotationEntry: KtAnnotationEntry, fileCache: FileCache): ClassId? =
-        fastResolveClassId(annotationEntry, fileCache)
-            ?: slowResolveClassId(annotationEntry)
-
-    /**
-     * Attempts to resolve the [ClassId] of an [annotationEntry] without full type resolution,
-     * by trying to guess the [ClassId] and asking the Analysis API if it is a type alias.
-     */
-    private fun fastResolveClassId(annotationEntry: KtAnnotationEntry, fileCache: FileCache): ClassId? {
-        val shortName = annotationEntry.shortName?.asString() ?: return null
-        val unresolvedTypeName = annotationEntry.typeReference?.text ?: return null
-
-        if (shortName != unresolvedTypeName) {
-            return null
-        }
-
-        val isPotentiallyShadowed = shortName in fileCache.localDeclarations
-        if (isPotentiallyShadowed) {
-            return null
-        }
-
-        // Check for explicit name import.
-        // If it exists, we have found the name, otherwise fall back to checking other candidates.
-        val candidateClassIds = fileCache.explicitImports[shortName]?.let { explicitFqn ->
-            // Direct name import found. Do not consider other imports.
-            listOf((ClassId.topLevel(explicitFqn)))
-        } ?: buildList {
-            // No direct name import found. Consider package imports, star imports, and default imports.
-            add(mkClassId(fileCache.packageName, shortName))
-            fileCache.starImports.forEach { add(mkClassId(it, shortName)) }
-            FileCache.DEFAULT_IMPORTS.forEach {
-                add(mkClassId(FqName(it), shortName))
-            }
-        }
-
-        return analyze {
-            candidateClassIds.firstNotNullOfOrNull { classId ->
-                when (val symbol = findClassLike(classId)) {
-                    null -> null
-
-                    is KaClassSymbol -> {
-                        // Class id found, return it!
-                        symbol.classId
-                    }
-
-                    is KaTypeAliasSymbol -> {
-                        // The symbol is a type alias. Fall back to expensive resolution
-                        symbol.expandedType.fullyExpandedType.symbol?.classId
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * Creates a [ClassId] from a package name and a top-level class name.
-     */
-    private fun mkClassId(packageFqName: FqName, topLevelName: String): ClassId =
-        ClassId(packageFqName, Name.identifier(topLevelName))
-
-    /**
-     * Resolves the [ClassId] of an [annotationEntry] using full Analysis API resolution.
-     *
-     * This is used as a fallback when [fastResolveClassId] cannot determine the class ID.
-     */
-    private fun slowResolveClassId(annotationEntry: KtAnnotationEntry): ClassId? = annotationEntry.classId
-
-    /**
-     * Resolves this [PsiElement] to the set of [KSAnnotated] symbols targeted by [annotation].
-     *
-     * The [annotation] is only required for Kotlin sources since annotations may have use site targets.
-     * Java sources never require an [annotation] to be present since annotations directly target the element
-     * being annotated.
-     */
-    private fun PsiElement.resolveToKSAnnotated(
-        annotation: KtAnnotationEntry? = null,
-    ): Collection<KSAnnotated> =
-        when (val element = this@resolveToKSAnnotated) {
-            // Kotlin sources
-            is KtDeclaration -> {
-                if (annotation == null) {
-                    throw InternalKSPException(
-                        "Unexpected null annotation",
-                        toLocation(),
-                        javaClass
-                    )
-                }
-                element.resolve(annotation)
-            }
-
-            is KtFile ->
-                listOf(element.resolve())
-
-            is KtTypeReference ->
-                emptyList() // Do nothing
-
-            // Java sources
-            is PsiParameter ->
-                listOf(element.resolve())
-
-            is PsiTypeParameter ->
-                element.resolve()
-
-            is PsiClass ->
-                listOf(element.resolve())
-
-            is PsiField -> element.resolveToProperty().let {
-                listOf(it, element.resolveToField(it))
-            }
-
-            is PsiMethod ->
-                listOf(element.resolve())
-
-            else ->
-                throw InternalKSPException(
-                    "Unreachable branch",
-                    element.toLocation(),
-                    element.javaClass
-                )
-        }
-
-    /**
-     * Resolves this [KtDeclaration] to the set of [KSAnnotated] symbols targeted by [annotationEntry].
-     */
-    private fun KtDeclaration.resolve(annotationEntry: KtAnnotationEntry): Collection<KSAnnotated> {
-        // TODO: This should perform case distinction instead of getTargetedSymbol
-        val ksSym = analyze { symbol.toKSAnnotated() }
-        return when (ksSym) {
-            is KSPropertyDeclaration -> targetFieldFix.resolveKSPropertyDeclaration(
-                ksSym,
-                annotationEntry
-            )
-
-            else -> ksSym.findTargetedSymbol(annotationEntry.ksUseSiteTarget)
-        }
-    }
-
-    /**
-     * Resolves the [KSFileImpl] corresponding to this [KtFile].
-     */
-    private fun KtFile.resolve(): KSFileImpl =
-        analyze { symbol }.toKSFile()
-
-    /**
-     * Resolves the [KSValueParameter] corresponding to this [PsiParameter].
-     */
-    private fun PsiParameter.resolve(): KSValueParameter {
-        val functionDecl = callableSymbol.toKSFunctionDeclaration()
-            ?: throw InternalKSPException(
-                "Failed to convert callable symbol to KSFunctionDeclaration",
-                toLocation(),
-                callableSymbol.javaClass
-            )
-        return functionDecl.parameters[parameterIndex()]
-    }
-
-    /**
-     * Resolves the [KSTypeParameter]s corresponding to this [PsiTypeParameter].
-     */
-    private fun PsiTypeParameter.resolve(): Collection<KSTypeParameter> = when (val decl = parent.parent) {
-        is PsiMethod ->
-            listOf(resolveTypeParameterOfMethod(decl))
-
-        is PsiClass ->
-            resolveTypeParameterOfClass(decl)
-
-        else ->
-            throw InternalKSPException(
-                "Unexpected Java declaration",
-                decl.toLocation(),
-                decl.javaClass
-            )
-    }
-
-    /**
-     * Resolves the [KSClassDeclarationImpl] corresponding to this [PsiClass].
-     */
-    private fun PsiClass.resolve(): KSClassDeclarationImpl {
-        val sym = analyze { namedClassSymbol }
-            ?: throw InternalKSPException(
-                "Unexpected null named class symbol",
-                toLocation(),
-                javaClass
-            )
-        return sym.toKSClassDeclaration()
-    }
-
-    /**
-     * Resolves the [KSPropertyDeclaration] corresponding to this [PsiField].
-     */
-    private fun PsiField.resolveToProperty(): KSPropertyDeclarationJavaImpl {
-        val sym = analyze { this@resolveToProperty.callableSymbol }
-            ?: throw InternalKSPException(
-                "Unexpected null callable symbol",
-                toLocation(),
-                javaClass
-            )
-        return sym.toKSPropertyDeclaration()
-            ?: throw InternalKSPException(
-                "Unexpected null KSPropertyDeclaration",
-                toLocation(),
-                javaClass
-            )
-    }
-
-    /**
-     * Resolves the [KSBackingField] corresponding to this [PsiField].
-     */
-    private fun PsiField.resolveToField(property: KSPropertyDeclarationJavaImpl? = null): KSBackingField {
-        val sym = analyze { this@resolveToField.callableSymbol }
-            ?: throw InternalKSPException(
-                "Unexpected null callable symbol",
-                toLocation(),
-                javaClass
-            )
-        throw InternalKSPException(
-            "Unexpected unreachable function. PsiField.resolveToField is not implemented yet",
-            toLocation(),
-            javaClass
-        )
-    }
-
-    /**
-     * Resolves the [KSFunctionDeclarationImpl] corresponding to this [PsiMethod].
-     */
-    private fun PsiMethod.resolve(): KSFunctionDeclarationImpl {
-        val sym = analyze { this@resolve.callableSymbol }
-            ?: throw InternalKSPException(
-                "Unexpected null callable symbol",
-                toLocation(),
-                javaClass
-            )
-        return sym.toKSFunctionDeclaration()
-            ?: throw InternalKSPException(
-                "Unexpected null KSFunctionDeclaration",
-                toLocation(),
-                javaClass
-            )
-    }
-
-    /**
-     * Returns the targeted symbols by a given annotated symbol at its use site target.
-     *
-     * E.g., `class A(@MyAnno val p: Int)` returns `p, p`, once for the parameter and once for the property.
-     */
-    private fun KSAnnotated.findTargetedSymbol(useSiteTarget: AnnotationUseSiteTarget?): Collection<KSAnnotated> =
-        when (useSiteTarget) {
-            null -> when (this) {
-                is KSValueParameter -> listOfNotNull(this, this.getGeneratedProperty())
-
-                is KSFunctionDeclarationImpl -> when {
-                    this.isGetter -> listOf(
-                        this.parentDeclaration?.getter()
-                            ?: throw InternalKSPException(
-                                "Unexpected missing getter",
-                                location,
-                                javaClass
-                            )
-                    )
-
-                    this.isSetter -> listOf(
-                        this.parentDeclaration?.setter()
-                            ?: throw InternalKSPException(
-                                "Unexpected missing setter",
-                                location,
-                                javaClass
-                            )
-                    )
-
-                    else -> listOf(this)
-                }
-
-                else -> listOf(this)
-            }
-
-            AnnotationUseSiteTarget.FILE -> when (this) {
-                is KSFile -> listOf(this)
-                else -> listOf(
-                    this.containingFile
-                        ?: throw InternalKSPException(
-                            "Unexpected missing file",
-                            location,
-                            javaClass
-                        )
-                )
-            }
-
-            AnnotationUseSiteTarget.PROPERTY -> when (this) {
-                is KSValueParameter -> listOf(
-                    this.getGeneratedProperty() ?: throw InternalKSPException(
-                        "Unexpected missing property for parameter",
-                        location,
-                        javaClass
-                    )
-                )
-
-                else -> listOf(this)
-            }
-
-            AnnotationUseSiteTarget.FIELD -> when (this) {
-                is KSValueParameter -> listOf(
-                    this.getGeneratedProperty()?.backingField ?: throw InternalKSPException(
-                        "Unexpected missing backing field for parameter",
-                        location,
-                        javaClass
-                    )
-                )
-
-                is KSPropertyDeclaration -> listOf(
-                    this.backingField ?: throw InternalKSPException(
-                        "Unexpected missing backing field for property",
-                        location,
-                        javaClass
-                    )
-                )
-
-                else -> listOf(this)
-            }
-
-            AnnotationUseSiteTarget.GET ->
-                listOf(
-                    this.getter()
-                        ?: throw InternalKSPException(
-                            "Unexpected missing getter",
-                            location,
-                            javaClass
-                        )
-                )
-
-            AnnotationUseSiteTarget.SET ->
-                listOf(
-                    this.setter()
-                        ?: throw InternalKSPException(
-                            "Unexpected missing setter",
-                            location,
-                            javaClass
-                        )
-                )
-
-            AnnotationUseSiteTarget.RECEIVER -> when (this) {
-                is KSFunctionDeclarationImpl -> listOf(
-                    extensionReceiver
-                        ?: throw InternalKSPException(
-                            "Unexpected missing extension receiver",
-                            location,
-                            javaClass
-                        )
-                )
-
-                is KSPropertyDeclaration -> listOf(
-                    extensionReceiver
-                        ?: throw InternalKSPException(
-                            "Unexpected missing extension receiver",
-                            location,
-                            javaClass
-                        )
-                )
-
-                else -> throw InternalKSPException(
-                    "Unexpected declaration",
-                    location,
-                    javaClass
-                )
-            }
-
-            AnnotationUseSiteTarget.PARAM -> when (this) {
-                is KSValueParameter -> listOf(this)
-                else -> throw InternalKSPException(
-                    "Unexpected annotated symbol with param use-site target",
-                    location,
-                    javaClass
-                )
-            }
-
-            AnnotationUseSiteTarget.SETPARAM -> when (this) {
-                is KSValueParameter -> {
-                    listOf(
-                        (this.parent as? KSFunctionDeclaration)?.parameters?.singleOrNull()
-                            ?: throw InternalKSPException(
-                                "Unexpected missing setter parameter",
-                                location,
-                                javaClass
-                            )
-                    )
-                }
-
-                is KSPropertyDeclaration -> listOf(
-                    setter?.parameter
-                        ?: throw InternalKSPException(
-                            "Unexpected missing setter parameter",
-                            location,
-                            javaClass
-                        )
-                )
-
-                else -> throw InternalKSPException(
-                    "Unexpected annotated symbol with 'setparam' use-site target",
-                    location,
-                    javaClass
-                )
-            }
-
-            AnnotationUseSiteTarget.DELEGATE ->
-                // TODO: Better impl for this
-                listOf(this)
-
-            AnnotationUseSiteTarget.ALL -> when (this) {
-                is KSValueParameter -> {
-                    buildList {
-                        add(this@findTargetedSymbol)
-                        this@findTargetedSymbol.getGeneratedProperty()?.findTargetedSymbol(useSiteTarget)
-                            ?.let { symbols ->
-                                symbols.forEach { addIfNotNull(it) }
-                            }
-                    }
-                }
-
-                is KSPropertyDeclaration ->
-                    // TODO: Add support for JvmRecord annotation according to KEEP 402
-                    buildList {
-                        add(this@findTargetedSymbol)
-                        addIfNotNull(this@findTargetedSymbol.getter)
-                        addIfNotNull(this@findTargetedSymbol.setter)
-                        addIfNotNull(this@findTargetedSymbol.setter?.parameter)
-                        addIfNotNull(this@findTargetedSymbol.backingField)
-                    }
-
-                else -> throw InternalKSPException(
-                    "Unexpected annotated symbol with 'all' use-site target",
-                    location,
-                    javaClass
-                )
-            }
-        }
-
-    /**
-     * Resolves this [PsiTypeParameter] to its [KSTypeParameter] representation within [method].
-     */
-    private fun PsiTypeParameter.resolveTypeParameterOfMethod(method: PsiMethod): KSTypeParameter {
-        val callableSym = analyze { method.callableSymbol }
-            ?: throw InternalKSPException(
-                "Unexpected null callable symbol",
-                method.toLocation(),
-                method.javaClass
-            )
-        val functionDecl = callableSym.toKSFunctionDeclaration()
-            ?: throw InternalKSPException(
-                "Failed to convert callable symbol to KSFunctionDeclaration",
-                method.toLocation(),
-                callableSym.javaClass
-            )
-        return functionDecl.typeParameters[parameterIndex]
-    }
-
-    /**
-     * Resolves this [PsiTypeParameter] to its [KSTypeParameter] representations within [clazz].
-     */
-    private fun PsiTypeParameter.resolveTypeParameterOfClass(clazz: PsiClass): List<KSTypeParameter> {
-        val classSym = analyze { clazz.namedClassSymbol }
-            ?: throw InternalKSPException(
-                "Unexpected named class symbol",
-                clazz.toLocation(),
-                clazz.javaClass
-            )
-        val typeParamSym = classSym.toKSClassDeclaration().typeParameters[parameterIndex]
-        // Type parameters for classes return two symbols with Analysis API,
-        // one as a Psi-based KaFirTypeParameter and one as a "regular" KaFirTypeParameter,
-        // so we return a clone here to maintain KSP API parity.
-        val clone = object : KSTypeParameter by typeParamSym {
-            override fun toString(): String = typeParamSym.toString()
-        }
-        return listOf(typeParamSym, clone)
-    }
-
-    /**
-     * Returns the short name used at the annotation site.
-     * Throws an exception if the name is `null`.
-     */
-    private val PsiAnnotation.shortName: String
-        get() = this.nameReferenceElement?.referenceName
-            ?: throw InternalKSPException(
-                "Unexpected nullable annotation short name",
-                toLocation(),
-                javaClass
-            )
-
-    /**
-     * Returns the callable symbol, i.e., the method symbol for the Java parameter `this`.
-     * Assumes that `this` is a parameter to a method.
-     */
-    private val PsiParameter.callableSymbol: KaCallableSymbol
-        get() {
-            val decl = this.parent.parent as? PsiMember
-                ?: throw InternalKSPException(
-                    "Unexpected PsiParameter",
-                    toLocation(),
-                    javaClass
-                )
-            return analyze {
-                decl.callableSymbol
-                    ?: throw InternalKSPException(
-                        "Unexpected null callable symbol",
-                        decl.toLocation(),
-                        decl.javaClass
-                    )
-            }
-        }
-
-    /**
-     * Returns the index of `this` in the list of type parameters.
-     */
-    private val PsiTypeParameter.parameterIndex: Int
-        get() {
-            when (val parent = this.parent) {
-                is PsiTypeParameterList ->
-                    return parent.getTypeParameterIndex(this)
-
-                else ->
-                    throw InternalKSPException(
-                        "Unexpected parent of PsiTypeParameter",
-                        parent.toLocation(),
-                        parent.javaClass
-                    )
-            }
-        }
-
-    /**
-     * The fully expanded [ClassId] of the annotation entry.
-     * This member is expensive to compute.
-     */
-    private val KtAnnotationEntry.classId: ClassId?
-        get() = analyze {
-            // N.B. do not use typeReference.type to get the ClassId because that can fail in certain edge cases, e.g.
-            //  https://github.com/google/ksp/issues/2913
-            this@classId.resolveToCall()
-                ?.successfulCallOrNull<KaAnnotationCall>()
-                ?.symbol
-                ?.containingClassId
-        }
-
-    /**
-     * The use-site target specified for this annotation entry.
-     */
-    private val KtAnnotationEntry.ksUseSiteTarget: AnnotationUseSiteTarget?
-        get() = useSiteTarget?.getAnnotationUseSiteTarget()?.toKSAnnotationUseSiteTarget()
-
-    /**
-     * This object represents a temporary fix for https://github.com/google/ksp/issues/2987.
-     * It decides if an annotation is annotated with the FIELD annotation target,
-     * and if so, applies to the field instead of the property.
-     */
-    private inner class AnnotationDeclarationTargetFieldFix {
-
-        /**
-         * Stores the [ClassId]s for the [KtAnnotationEntry].
-         */
-        private val annotationClassIds: MutableMap<KtAnnotationEntry, ClassId> = mutableMapOf()
-
-        /**
-         * Stores the key-value pair [annotation] -> [classId]. This mapping is used by [resolveKSPropertyDeclaration].
-         */
-        fun recordClassId(annotation: KtAnnotationEntry, classId: ClassId) {
-            annotationClassIds.putIfAbsent(annotation, classId)
-        }
-
-        /**
-         * Resolves [ksPropDecl] with respect to the declaration-site target [annotation].
-         * The use-site target of [annotation] takes precedence over the declaration-site target.
-         */
-        fun resolveKSPropertyDeclaration(
-            ksPropDecl: KSPropertyDeclaration,
-            annotation: KtAnnotationEntry
-        ): Collection<KSAnnotated> =
-            when {
-                annotation.ksUseSiteTarget != null -> ksPropDecl.findTargetedSymbol(annotation.ksUseSiteTarget)
-
-                canBeAppliedTo(annotation, ANNOTATION_TARGET_PROPERTY_ENTRY) -> {
-                    listOfNotNull(ksPropDecl)
-                }
-
-                canBeAppliedTo(annotation, ANNOTATION_TARGET_FIELD_ENTRY) -> {
-                    listOfNotNull(ksPropDecl.backingField)
-                }
-
-                else -> emptyList()
-            }
-
-        /**
-         * Returns `true` if [annotation] can be applied to [entry], which is a [CallableId] for
-         * the FIELD or PROPERTY [AnnotationTarget].
-         * @see [ANNOTATION_TARGET_FIELD_ENTRY]
-         * @see [ANNOTATION_TARGET_PROPERTY_ENTRY]
-         */
-        private fun canBeAppliedTo(annotation: KtAnnotationEntry, entry: CallableId): Boolean =
-            applicationCache.getOrPut(annotation to entry) {
-                val targetAnnotation = findTargetAnnotation(getCachedClassId(annotation)) ?: return@getOrPut true
-                val args = targetAnnotationArguments(targetAnnotation)
-                args.contains(entry)
-            }
-
-        /**
-         * Private cache for [canBeAppliedTo]
-         */
-        private val applicationCache: MutableMap<Pair<KtAnnotationEntry, CallableId>, Boolean> = mutableMapOf()
-
-        private fun getCachedClassId(annotation: KtAnnotationEntry): ClassId =
-            annotationClassIds[annotation] ?: throw InternalKSPException(
-                "Unexpected null class id for annotation entry: $annotation",
-                annotation.psiOrParent.toLocation(),
-                annotation.javaClass
-            )
-
-        private val TARGET_ANNOTATION = ClassId(
-            FqName("kotlin.annotation"),
-            Name.identifier("Target")
-        )
-        private val ANNOTATION_TARGET_FIELD_ENTRY: CallableId = annotationTargetEntry("FIELD")
-        private val ANNOTATION_TARGET_PROPERTY_ENTRY: CallableId = annotationTargetEntry("PROPERTY")
-
-        private fun annotationTargetEntry(entry: String) = ClassId(
-            FqName("kotlin.annotation"),
-            Name.identifier("AnnotationTarget")
-        ).let {
-            CallableId(it, Name.identifier(entry))
-        }
-
-        private fun isAnnotationClass(kaClassSym: KaClassSymbol?): Boolean =
-            kaClassSym?.classKind == KaClassKind.ANNOTATION_CLASS
-
-        /**
-         * Returns the `@Target` annotation if it exists.
-         */
-        private fun findTargetAnnotation(kaClassSym: KaClassSymbol?): KaAnnotation? {
-            if (!isAnnotationClass(kaClassSym)) {
-                return null
-            }
-            return kaClassSym?.annotations?.find { it.classId == TARGET_ANNOTATION }
-        }
-
-        private fun findTargetAnnotation(classId: ClassId): KaAnnotation? =
-            findTargetAnnotation(classId.toKtClassSymbol())
-
-        /**
-         * Returns the arguments of the target annotation. Throws an error if the annotation
-         * has more than one non-null argument.
-         */
-        private fun targetAnnotationArguments(targetAnnotation: KaAnnotation): List<CallableId> {
-            val args = targetAnnotation.arguments
-            if (args.size != 1) {
-                throw InternalKSPException(
-                    "Unexpected number of arguments passed to @Target annotation",
-                    targetAnnotation.psi.toLocation(),
-                    targetAnnotation.javaClass
-                )
-            }
-            return args.firstOrNull()?.let { flattenAnnotationArgs(it.expression) } ?: emptyList()
-        }
-
-        /**
-         * Returns a list of [CallableId] where each member is the concrete type of the argument passed to
-         * the [Target] annotation.
-         * The [Target] annotation is defined as:
-         * ```kt
-         * annotation class Target(val allowedTarget: AnnotationTarget)
-         * ```
-         * However, you can pass an array of [AnnotationTarget] to [Target], so
-         * both of the following constructors are allowed, hence the flattening:
-         * ```kt
-         * @Target(AnnotationTarget.FIELD)
-         * @Target({AnnotationTarget.FIELD, AnnotationTarget.CLASS})
-         * ```
-         */
-        private fun flattenAnnotationArgs(annotationValue: KaAnnotationValue): List<CallableId> =
-            when (annotationValue) {
-                is KaAnnotationValue.EnumEntryValue -> listOfNotNull(annotationValue.callableId)
-                is KaAnnotationValue.ArrayValue -> annotationValue.values.flatMap(::flattenAnnotationArgs)
-                else -> throw InternalKSPException(
-                    "Unexpected argument type to @Target annotation",
-                    annotationValue.sourcePsi.toLocation(),
-                    annotationValue.javaClass,
-                )
-            }
-    }
-}
